story-service/requirements.txt
fastapi==0.115.5
instaloader==4.14.2
python-dotenv==1.0.1
uvicorn==0.32.0
